btrfs-progs: add HMAC(SHA256) support
Add support for authenticated file systems using HMAC(SHA256) as
checksumming algorithm.

Example:
  $ mkfs.btrfs --csum hmac-sha256 --auth-key 0123456789 -f test.img

Note: this is not final format, the key is specified as raw bytes but
must be ascii without a null. In the future the key will be retrieved
eg. from keyrings

Note 2: the spec string of the authentication will change.

Note 3: there's no builtin fallback implementation for crypto, so
either libsodium or libgcrypt is now mandatory, this will be fixed.

Signed-off-by: Johannes Thumshirn <johannes.thumshirn@wdc.com>
[ add notes ]
Signed-off-by: David Sterba <dsterba@suse.com>