btrfs-progs: crypto: add auth key API
Add API for authentication key to provide a wider range of ways how to
specify it on the command line and make it ready for use. As many
commands would need to access the key to at least verify the checksums,
each command needs to take a new parameter. To avoid adding several
parameters for different formats, use specifier that can be extended as
needed.

Currently there's only raw bytes format implemented. The planned list
is:

 - id:1234
 - name:keyname
 - file:/read/from/this/file
 - fd:3
 - raw:useonlyfortesting

The first two (id and name) are identifiers for keyctl/keyutils. The
other formats are for convenience and for testing. Only one format type
can be present in the spec buffer.

The API provides:

- auth_key_init  - initialize buffer
- auth_key_reset - free any existing data of the given type and
                   initialize the buffer
- auth_key_parse - read the string spec, validate it and extract
		   information that can be used to properly setup the
		   key, this does not read nor setup the key bytes
- auth_key_setup - read the key bytes from the source and store it into
                   the spec buffer

Signed-off-by: David Sterba <dsterba@suse.com>