btrfs-progs: crypto: convert auth key to the API
Convert raw buffer to the auth key spec API. This now supports the
format in mkfs, but nothing else yet. Other utilities will refuse t
open the filesystem, 'inspect dump-super' will show that the checksum
mismatches.

The output of mkfs will show the specifier as-is, this means that the
raw bytes are shown as well and that's why this is just for testing.

  $ mkfs.btrfs --csum auth-sha256 --auth-key raw:1234 img
  Label:              (null)
  UUID:               2eb8181c-3288-4769-a74f-2dbda2e050aa
  Node size:          16384
  Sector size:        4096
  Filesystem size:    1.00GiB
  Block group profiles:
    Data:             single            8.00MiB
    Metadata:         DUP              51.19MiB
    System:           DUP               8.00MiB
  SSD detected:       no
  Zoned device:       no
  Incompat features:  extref, skinny-metadata
  Runtime features:
  Checksum:           auth-sha256
    Auth key:         raw:1234
  Number of devices:  1
  Devices:
     ID        SIZE  PATH
      1     1.00GiB  img

Signed-off-by: David Sterba <dsterba@suse.com>